Updated code

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	long int i, j = 1, k = 2, next;
+	int i, j = 1, k = 2, next;
 
 	for (i = 1; i <= 50; i++)
 	{
